Introduce new suite of types for polaris web components
