+++ packages/ui-extensions/src/surfaces/point-of-sale/components/Navigator.d.ts
@@ -1,50 +1,43 @@
-/** VERSION: 0.0.0 **/
+/* VERSION: latest */
 /* eslint-disable import/extensions */
 /* eslint-disable @typescript-eslint/no-namespace */
-/* eslint-disable @typescript-eslint/member-ordering */
-
 // eslint-disable-next-line @typescript-eslint/triple-slash-reference, spaced-comment
 /// <reference lib="DOM" />
-import type {
-  PreactBaseElementPropsWithChildren,
-  PreactCustomElement,
-} from './shared.d.ts';
+import type {Key, Ref, ComponentChildren} from './components-shared.d.ts';
 
 /**
- * A component used to navigate between different screens.
+ * Used when an element does not have children.
  */
-export interface NavigatorProps {
-  /**
-   * The name of the screen to show initially.
-   */
+export interface BaseElementProps<TClass = HTMLElement> {
+  key?: Key;
+  ref?: Ref<TClass>;
+  slot?: Lowercase<string>;
+}
+/**
+ * Used when an element has children.
+ */
+export interface BaseElementPropsWithChildren<TClass = HTMLElement>
+  extends BaseElementProps<TClass> {
+  children?: ComponentChildren;
+}
+
+declare const tagName = 's-navigator';
+export interface NavigatorJSXProps {
   initialScreenName?: string;
+  children?: React.ReactNode;
 }
-
-declare class Navigator extends PreactCustomElement implements NavigatorProps {
-  accessor initialScreenName: NavigatorProps['initialScreenName'];
-  constructor();
-}
-
 declare global {
   interface HTMLElementTagNameMap {
-    [tagName]: Navigator;
+    [tagName]: NavigatorJSXProps;
   }
 }
-
 declare module 'preact' {
   namespace createElement.JSX {
     interface IntrinsicElements {
-      [tagName]: NavigatorJSXProps &
-        PreactBaseElementPropsWithChildren<Navigator>;
+      [tagName]: BaseElementPropsWithChildren<NavigatorJSXProps>;
     }
   }
 }
 
-declare const tagName = 's-navigator';
-
-export interface NavigatorJSXProps extends Partial<NavigatorProps> {
-  id?: string;
-}
-
-export {Navigator};
+export {tagName};
 export type {NavigatorJSXProps};
