+++ packages/ui-extensions/src/surfaces/point-of-sale/components/Text.d.ts
@@ -1,68 +1,55 @@
-/** VERSION: 0.0.0 **/
+/* VERSION: latest */
 /* eslint-disable import/extensions */
 /* eslint-disable @typescript-eslint/no-namespace */
-/* eslint-disable @typescript-eslint/member-ordering */
-
 // eslint-disable-next-line @typescript-eslint/triple-slash-reference, spaced-comment
 /// <reference lib="DOM" />
 import type {
-  PreactBaseElementPropsWithChildren,
-  PreactCustomElement,
-} from './shared.d.ts';
+  TextProps,
+  ComponentChildren$1,
+  Key,
+  Ref,
+} from './components-shared.d.ts';
 
-export type TextVariant =
-  | 'heading1'
-  | 'heading2'
-  | 'heading3'
-  | 'body'
-  | 'caption1'
-  | 'caption2';
-
-export type ColorType =
-  | 'TextPrimary'
-  | 'TextSecondary'
-  | 'TextCritical'
-  | 'TextSuccess';
-
+type ComponentChildren = ComponentChildren$1;
 /**
- * Text can be rendered in different sizes and colors in order to structure content.
+ * Used when an element does not have children.
  */
-export interface TextProps {
-  /**
-   * The visual style variant of the text.
-   */
-  variant?: TextVariant;
-  /**
-   * The color of the text.
-   */
-  color?: ColorType;
+export interface BaseElementProps<TClass = HTMLElement> {
+  key?: Key;
+  ref?: Ref<TClass>;
+  slot?: Lowercase<string>;
+}
+/**
+ * Used when an element has children.
+ */
+export interface BaseElementPropsWithChildren<TClass = HTMLElement>
+  extends BaseElementProps<TClass> {
+  children?: ComponentChildren$1;
 }
 
-declare class Text extends PreactCustomElement implements TextProps {
-  accessor variant: TextProps['variant'];
-  accessor color: TextProps['color'];
-  constructor();
+declare const tagName = 's-text';
+type AlignedProps = Pick<TextProps, 'tone' | 'type'>;
+export interface TextJSXProps extends AlignedProps {
+  color?: 'subdued' | 'base' | 'strong';
+  type?: Extract<TextProps['type'], 'emphasis' | 'strong' | 'small' | 'offset'>;
+  tone?: Extract<
+    TextProps['tone'],
+    'auto' | 'neutral' | 'info' | 'success' | 'warning' | 'critical'
+  >;
+  children?: ComponentChildren;
 }
-
 declare global {
   interface HTMLElementTagNameMap {
-    [tagName]: Text;
+    [tagName]: TextJSXProps;
   }
 }
-
 declare module 'preact' {
   namespace createElement.JSX {
     interface IntrinsicElements {
-      [tagName]: TextJSXProps & PreactBaseElementPropsWithChildren<Text>;
+      [tagName]: BaseElementPropsWithChildren<TextJSXProps>;
     }
   }
 }
 
-declare const tagName = 's-text';
-
-export interface TextJSXProps extends Partial<TextProps> {
-  id?: string;
-}
-
-export {Text};
+export {tagName};
 export type {TextJSXProps};
