+++ packages/ui-extensions/src/surfaces/point-of-sale/components/TextField.d.ts
@@ -1,88 +1,61 @@
-/** VERSION: 0.0.0 **/
+/* VERSION: latest */
 /* eslint-disable import/extensions */
 /* eslint-disable @typescript-eslint/no-namespace */
-/* eslint-disable @typescript-eslint/member-ordering */
-
 // eslint-disable-next-line @typescript-eslint/triple-slash-reference, spaced-comment
 /// <reference lib="DOM" />
 import type {
-  CallbackEvent,
-  CallbackEventListener,
-  PreactBaseElementPropsWithChildren,
-  PreactCustomElement,
-} from './shared.d.ts';
+  TextFieldProps,
+  Key,
+  Ref,
+  ComponentChildren,
+} from './components-shared.d.ts';
 
 /**
- * Use a text field to allow merchants to enter or edit text.
+ * Used when an element does not have children.
  */
-export interface TextFieldProps {
-  /**
-   * The label displayed above the text field.
-   */
-  label?: string;
-  /**
-   * The placeholder text shown when the field is empty.
-   */
-  placeholder?: string;
-  /**
-   * Whether the field is required.
-   */
-  required?: boolean;
-  /**
-   * The current value of the text field.
-   */
-  value?: string;
-  /**
-   * Callback executed when the text changes.
-   */
-  onChange?: (value: string) => void;
-  /**
-   * Whether the field is editable.
-   */
-  editable?: boolean;
-  /**
-   * Whether the field should obscure text (for passwords).
-   */
-  secure?: boolean;
-  /**
-   * The type of keyboard to display.
-   */
-  keyboardType?: 'default' | 'email' | 'numeric' | 'phone';
+export interface BaseElementProps<TClass = HTMLElement> {
+  key?: Key;
+  ref?: Ref<TClass>;
+  slot?: Lowercase<string>;
 }
-
-declare class TextField extends PreactCustomElement implements TextFieldProps {
-  accessor label: TextFieldProps['label'];
-  accessor placeholder: TextFieldProps['placeholder'];
-  accessor required: TextFieldProps['required'];
-  accessor value: TextFieldProps['value'];
-  accessor onchange: CallbackEventListener<typeof tagName> | null;
-  accessor editable: TextFieldProps['editable'];
-  accessor secure: TextFieldProps['secure'];
-  accessor keyboardType: TextFieldProps['keyboardType'];
-  constructor();
+/**
+ * Used when an element has children.
+ */
+export interface BaseElementPropsWithChildren<TClass = HTMLElement>
+  extends BaseElementProps<TClass> {
+  children?: ComponentChildren;
 }
+type CallbackEvent<
+  TTagName extends keyof HTMLElementTagNameMap,
+  TEvent extends Event = Event,
+> = TEvent & {
+  currentTarget: HTMLElementTagNameMap[TTagName];
+};
 
+declare const tagName = 's-text-field';
+type AlignedProps = Pick<
+  TextFieldProps,
+  'label' | 'details' | 'value' | 'placeholder'
+>;
+export interface TextFieldEventProps {
+  onInput?: ((event: CallbackEvent<typeof tagName>) => void) | null;
+  onFocus?: ((event: CallbackEvent<typeof tagName>) => void) | null;
+  onBlur?: ((event: CallbackEvent<typeof tagName>) => void) | null;
+  onChange?: ((event: CallbackEvent<typeof tagName>) => void) | null;
+}
+export interface TextFieldJSXProps extends AlignedProps, TextFieldEventProps {}
 declare global {
   interface HTMLElementTagNameMap {
-    [tagName]: TextField;
+    [tagName]: TextFieldJSXProps;
   }
 }
-
 declare module 'preact' {
   namespace createElement.JSX {
     interface IntrinsicElements {
-      [tagName]: TextFieldJSXProps &
-        PreactBaseElementPropsWithChildren<TextField>;
+      [tagName]: BaseElementPropsWithChildren<TextFieldJSXProps>;
     }
   }
 }
 
-declare const tagName = 's-text-field';
-
-export interface TextFieldJSXProps extends Partial<TextFieldProps> {
-  onChange?: ((event: CallbackEvent<typeof tagName>) => void) | null;
-  id?: string;
-}
-
-export {TextField};
-export type {TextFieldJSXProps};
+export {tagName};
+export type {TextFieldEventProps, TextFieldJSXProps};
