+++ packages/ui-extensions/src/surfaces/point-of-sale/components/components-shared.d.ts
@@ -1,8 +1,6 @@
 /* VERSION: latest */
 /* eslint-disable import/extensions */
 /* eslint-disable @typescript-eslint/no-namespace */
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference, spaced-comment
-/// <reference lib="DOM" />
 /**
  * TODO: Update `any` type here after this is resolved
  * https://github.com/Shopify/ui-api-design/issues/139
@@ -2242,761 +2240,4 @@ export interface Context<T> extends Provider<T> {
   Consumer: Consumer<T>;
   Provider: Provider<T>;
   displayName?: string;
-}
-
-type ComponentChildren = ComponentChildren$1;
-type ComponentChild = ComponentChild$1;
-/**
- * Used when an element does not have children.
- */
-interface BaseElementProps<TClass = HTMLElement> {
-  key?: Key;
-  ref?: Ref<TClass>;
-  slot?: Lowercase<string>;
-}
-/**
- * Used when an element has children.
- */
-interface BaseElementPropsWithChildren<TClass = HTMLElement>
-  extends BaseElementProps<TClass> {
-  children?: ComponentChildren$1;
-}
-type CallbackEvent<
-  TTagName extends keyof HTMLElementTagNameMap,
-  TEvent extends Event = Event,
-> = TEvent & {
-  currentTarget: HTMLElementTagNameMap[TTagName];
-};
-type CallbackToggleEvent<
-  TTagName extends keyof HTMLElementTagNameMap,
-  TEvent extends ToggleEvent = ToggleEvent,
-> = TEvent & {
-  currentTarget: HTMLElementTagNameMap[TTagName];
-};
-
-declare const tagName$e = 's-button';
-type AlignedProps$b = Pick<
-  ButtonProps,
-  | 'accessibilityLabel'
-  | 'disabled'
-  | 'command'
-  | 'commandFor'
-  | 'loading'
-  | 'tone'
-  | 'variant'
-  | 'id'
->;
-interface ButtonEventProps {
-  onClick?: (event: CallbackEvent<typeof tagName$e>) => void;
-}
-interface ButtonJSXProps extends AlignedProps$b, ButtonEventProps {
-  tone?: Extract<ButtonProps['tone'], 'auto' | 'critical'>;
-  variant?: Extract<ButtonProps['variant'], 'primary' | 'secondary'>;
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$e]: ButtonJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$e]: BaseElementPropsWithChildren<ButtonJSXProps>;
-    }
-  }
-}
-
-declare const tagName$d = 's-text';
-type AlignedProps$a = Pick<TextProps, 'tone' | 'type'>;
-interface TextJSXProps extends AlignedProps$a {
-  color?: 'subdued' | 'base' | 'strong';
-  type?: Extract<TextProps['type'], 'emphasis' | 'strong' | 'small' | 'offset'>;
-  tone?: Extract<
-    TextProps['tone'],
-    'auto' | 'neutral' | 'info' | 'success' | 'warning' | 'critical'
-  >;
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$d]: TextJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$d]: BaseElementPropsWithChildren<TextJSXProps>;
-    }
-  }
-}
-
-declare const tagName$c = 's-scroll-box';
-type AlignedProps$9 = Pick<ScrollBoxProps, 'id'>;
-interface ScrollBoxJSXProps extends AlignedProps$9 {
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$c]: ScrollBoxJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$c]: BaseElementPropsWithChildren<ScrollBoxJSXProps>;
-    }
-  }
-}
-
-declare const tagName$b = 's-tile';
-interface TileEventProps {
-  onPress?: () => void;
-}
-interface TileJSXProps extends TileEventProps {
-  title: string;
-  subtitle?: string;
-  enabled?: boolean;
-  destructive?: boolean;
-  badgeValue?: string;
-  testID?: string;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$b]: TileJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$b]: BaseElementPropsWithChildren<TileJSXProps>;
-    }
-  }
-}
-
-declare const tagName$a = 's-navigator';
-interface NavigatorJSXProps {
-  initialScreenName?: string;
-  children?: React.ReactNode;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$a]: NavigatorJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$a]: BaseElementPropsWithChildren<NavigatorJSXProps>;
-    }
-  }
-}
-
-declare const tagName$9 = 's-screen';
-interface ScreenSlotProps {
-  /**
-   * Displays a secondary action button on the screen.
-   */
-  secondaryActions?: ComponentChild;
-}
-interface ScreenEventProps {
-  /**
-   * Triggered when the screen is navigated to.
-   */
-  onNavigate?: () => void;
-  /**
-   * Triggered when the user navigates back from this screen. Runs after screen is unmounted.
-   */
-  onNavigateBack?: () => void;
-  /**
-   * A callback that gets triggered when the navigation event completes and the screen receives the parameters.
-   */
-  onReceiveParams?: (params: any) => void;
-}
-interface ScreenJSXProps extends ScreenSlotProps, ScreenEventProps {
-  /**
-   * Used to identify this screen as a destination in the navigation stack.
-   */
-  name: string;
-  /**
-   * The title of the screen which will be displayed on the UI.
-   */
-  title: string;
-  /**
-   * Displays a loading indicator when `true`.
-   * Set this to `true` when performing an asynchronous task, and then to false when the data becomes available to the UI.
-   */
-  isLoading?: boolean;
-  /**
-   * Dictates how the `Screen` will be presented when navigated to.
-   */
-  presentation?: ScreenPresentationProps;
-  children?: React.ReactNode;
-}
-/** Represents the presentation of a screen in the navigation stack.
- * @property `sheet` displays the screen from the bottom on `navigate` when `true`.
- */
-interface ScreenPresentationProps {
-  /**
-   * Displays the screen from the bottom on `navigate` when `true`.
-   */
-  sheet?: boolean;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$9]: ScreenJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$9]: BaseElementPropsWithChildren<ScreenJSXProps>;
-    }
-  }
-}
-
-declare const tagName$8 = 's-banner';
-type AlignedProps$8 = Pick<BannerProps, 'heading' | 'hidden' | 'tone' | 'id'>;
-interface BannerSlotProps {
-  primaryAction?: ComponentChild;
-}
-interface BannerJSXProps extends AlignedProps$8, BannerSlotProps {
-  tone?: Extract<
-    BannerProps['tone'],
-    'success' | 'info' | 'warning' | 'critical'
-  >;
-  primaryAction?: ComponentChild;
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$8]: BannerJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$8]: BaseElementPropsWithChildren<
-        Omit<BannerJSXProps, 'primaryAction'>
-      >;
-    }
-  }
-}
-
-declare const tagName$7 = 's-box';
-type AlignedProps$7 = Pick<
-  BoxProps,
-  | 'id'
-  | 'padding'
-  | 'paddingBlock'
-  | 'paddingBlockStart'
-  | 'paddingBlockEnd'
-  | 'paddingInline'
-  | 'paddingInlineStart'
-  | 'paddingInlineEnd'
-  | 'blockSize'
-  | 'minBlockSize'
-  | 'maxBlockSize'
-  | 'inlineSize'
-  | 'minInlineSize'
-  | 'maxInlineSize'
-  | 'accessibilityRole'
-  | 'accessibilityLabel'
-  | 'border'
-  | 'borderColor'
-  | 'borderRadius'
-  | 'borderWidth'
-  | 'background'
->;
-interface BoxJSXProps extends AlignedProps$7 {
-  children?: ComponentChildren;
-  overflow?: Extract<BoxProps['overflow'], 'hidden' | 'visible'>;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$7]: BoxJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$7]: BaseElementPropsWithChildren<BoxJSXProps>;
-    }
-  }
-}
-
-declare const tagName$6 = 's-icon';
-type SupportedIconNames = Extract<
-  IconProps['type'],
-  | 'arrow-down'
-  | 'arrow-left'
-  | 'arrow-right'
-  | 'arrow-up'
-  | 'apps'
-  | 'analytics'
-  | 'backspace'
-  | 'call'
-  | 'cancel'
-  | 'cash'
-  | 'checkmark'
-  | 'cart'
-  | 'cart-filled'
-  | 'check'
-  | 'check-circle'
-  | 'alert-circle'
-  | 'info'
-  | 'phone'
-  | 'edit'
-  | 'delete'
-  | 'settings'
-  | 'search'
-  | 'print'
-  | 'refresh'
-  | 'help'
-  | 'plus'
-  | 'minus'
-  | 'email'
-  | 'gift-card'
-  | 'discount'
-  | 'receipt'
-  | 'chevron-down'
-  | 'chevron-left'
-  | 'chevron-right'
-  | 'chevron-up'
-  | 'menu'
-  | 'home'
-  | 'location'
-  | 'time'
-  | 'calendar'
-  | 'products'
-  | 'orders'
-  | 'customers'
-  | 'collection'
->;
-type SupportedSizes = Extract<
-  IconProps['size'],
-  'small' | 'base' | 'large' | 'large-100'
->;
-type AlignedProps$6 = Pick<IconProps, 'tone' | 'color'>;
-interface IconJSXProps extends AlignedProps$6 {
-  /**
-   * The type of icon to display. Maps to PDS icon names.
-   * @default ''
-   */
-  type?: SupportedIconNames;
-  /**
-   * The size of the icon.
-   * @default 'base'
-   */
-  size?: SupportedSizes;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$6]: IconJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$6]: BaseElementProps<IconJSXProps>;
-    }
-  }
-}
-
-declare const tagName$5 = 's-stack';
-type AlignedProps$5 = Pick<
-  StackProps,
-  | 'id'
-  | 'accessibilityLabel'
-  | 'alignItems'
-  | 'alignContent'
-  | 'gap'
-  | 'background'
-  | 'blockSize'
-  | 'border'
-  | 'borderColor'
-  | 'borderRadius'
-  | 'borderStyle'
-  | 'borderWidth'
-  | 'columnGap'
-  | 'direction'
-  | 'inlineSize'
-  | 'justifyContent'
-  | 'maxBlockSize'
-  | 'maxInlineSize'
-  | 'minBlockSize'
-  | 'minInlineSize'
-  | 'overflow'
-  | 'padding'
-  | 'paddingBlock'
-  | 'paddingBlockStart'
-  | 'paddingBlockEnd'
-  | 'paddingInline'
-  | 'paddingInlineStart'
-  | 'paddingInlineEnd'
-  | 'rowGap'
->;
-interface StackJSXProps extends AlignedProps$5 {
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$5]: StackJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$5]: BaseElementPropsWithChildren<StackJSXProps>;
-    }
-  }
-}
-
-declare const tagName$4 = 's-badge';
-type AlignedProps$4 = Pick<BadgeProps, 'id'>;
-interface BadgeJSXProps extends AlignedProps$4 {
-  tone?: Extract<
-    BadgeProps['tone'],
-    'auto' | 'neutral' | 'info' | 'success' | 'warning' | 'critical'
-  >;
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$4]: BadgeJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$4]: BaseElementPropsWithChildren<BadgeJSXProps>;
-    }
-  }
-}
-
-declare const tagName$3 = 's-choice-list';
-type AlignedProps$3 = Pick<ChoiceListProps, 'values' | 'multiple'>;
-interface ChoiceListEventProps {
-  onChange?: ((event: CallbackEvent<typeof tagName$3>) => void) | null;
-  onInput?: ((event: CallbackEvent<typeof tagName$3>) => void) | null;
-}
-interface ChoiceListJSXProps extends AlignedProps$3, ChoiceListEventProps {
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$3]: ChoiceListJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$3]: BaseElementPropsWithChildren<ChoiceListJSXProps>;
-    }
-  }
-}
-
-declare const tagName$2 = 's-choice';
-type AlignedProps$2 = Pick<ChoiceProps, 'value' | 'disabled' | 'selected'>;
-interface ChoiceJSXProps extends AlignedProps$2 {
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$2]: ChoiceJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$2]: BaseElementPropsWithChildren<ChoiceJSXProps>;
-    }
-  }
-}
-
-declare const tagName$1 = 's-modal';
-type AlignedProps$1 = Pick<ModalProps, 'id'>;
-interface ModalEventProps {
-  onHide?: (event: CallbackEvent<typeof tagName$1>) => void | null;
-  onShow?: (event: CallbackEvent<typeof tagName$1>) => void | null;
-  onAfterHide?: (event: CallbackEvent<typeof tagName$1>) => void | null;
-  onAfterShow?: (event: CallbackEvent<typeof tagName$1>) => void | null;
-  onToggle?: (event: CallbackToggleEvent<typeof tagName$1>) => void | null;
-  onAfterToggle?: (event: CallbackToggleEvent<typeof tagName$1>) => void | null;
-}
-interface ModalJSXProps extends AlignedProps$1, ModalEventProps {
-  children?: ComponentChildren;
-}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName$1]: ModalJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName$1]: BaseElementPropsWithChildren<ModalJSXProps>;
-    }
-  }
-}
-
-declare const tagName = 's-text-field';
-type AlignedProps = Pick<
-  TextFieldProps,
-  'label' | 'details' | 'value' | 'placeholder'
->;
-interface TextFieldEventProps {
-  onInput?: ((event: CallbackEvent<typeof tagName>) => void) | null;
-  onFocus?: ((event: CallbackEvent<typeof tagName>) => void) | null;
-  onBlur?: ((event: CallbackEvent<typeof tagName>) => void) | null;
-  onChange?: ((event: CallbackEvent<typeof tagName>) => void) | null;
-}
-interface TextFieldJSXProps extends AlignedProps, TextFieldEventProps {}
-declare global {
-  interface HTMLElementTagNameMap {
-    [tagName]: TextFieldJSXProps;
-  }
-}
-declare module 'preact' {
-  namespace createElement.JSX {
-    interface IntrinsicElements {
-      [tagName]: BaseElementPropsWithChildren<TextFieldJSXProps>;
-    }
-  }
-}
-
-export type {
-  BadgeJSXProps,
-  BannerJSXProps,
-  BoxJSXProps,
-  ButtonJSXProps,
-  ChoiceJSXProps,
-  ChoiceListJSXProps,
-  IconJSXProps,
-  ModalJSXProps,
-  NavigatorJSXProps,
-  ScreenJSXProps,
-  ScrollBoxJSXProps,
-  StackJSXProps,
-  TextFieldJSXProps,
-  TextJSXProps,
-  TileJSXProps,
-};
-
-interface BannerSlots {
-  'primary-action'?: HTMLElement;
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$e]: BaseElementPropsWithChildren<ButtonJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$e]: BaseElementPropsWithChildren<ButtonJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$d]: BaseElementPropsWithChildren<TextJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$d]: BaseElementPropsWithChildren<TextJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$c]: BaseElementPropsWithChildren<ScrollBoxJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$c]: BaseElementPropsWithChildren<ScrollBoxJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$b]: BaseElementPropsWithChildren<TileJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$b]: BaseElementPropsWithChildren<TileJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$a]: BaseElementPropsWithChildren<NavigatorJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$a]: BaseElementPropsWithChildren<NavigatorJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$9]: BaseElementPropsWithChildren<ScreenJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$9]: BaseElementPropsWithChildren<ScreenJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$8]: BaseElementPropsWithChildren<
-        Omit<BannerJSXProps, 'primaryAction'>
-      >;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$8]: BaseElementPropsWithChildren<
-        Omit<BannerJSXProps, 'primaryAction'>
-      >;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$7]: BaseElementPropsWithChildren<BoxJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$7]: BaseElementPropsWithChildren<BoxJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$6]: BaseElementProps<IconJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$6]: BaseElementProps<IconJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$5]: BaseElementPropsWithChildren<StackJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$5]: BaseElementPropsWithChildren<StackJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$4]: BaseElementPropsWithChildren<BadgeJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$4]: BaseElementPropsWithChildren<BadgeJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$3]: BaseElementPropsWithChildren<ChoiceListJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$3]: BaseElementPropsWithChildren<ChoiceListJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$2]: BaseElementPropsWithChildren<ChoiceJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$2]: BaseElementPropsWithChildren<ChoiceJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$1]: BaseElementPropsWithChildren<ModalJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName$1]: BaseElementPropsWithChildren<ModalJSXProps>;
-    }
-  }
-}
-declare module 'react' {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName]: BaseElementPropsWithChildren<TextFieldJSXProps>;
-    }
-  }
-}
-declare global {
-  namespace JSX {
-    interface IntrinsicElements {
-      [tagName]: BaseElementPropsWithChildren<TextFieldJSXProps>;
-    }
-  }
 }
